css-text/word-break/word-break-break-all-008.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8">
<title>word-break: break-all, syllabic cluster</title>
<meta name="assert" content="[Exploratory test] word-break: break-all doesn't create break opportunities between the components of an Indic conjunct cluster.">
<link rel='help' href='https://drafts.csswg.org/css-text-3/#word-break-property'>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<style type='text/css'>
.test { word-break: break-all; }
/* the CSS below is not part of the test */
.test, .ref { margin: 20px;  font: 3em/1 sans-serif; position:absolute; top: 0; width:90%;}
.ref { color:red; z-index: -100; }
</style>
</head>
<body>
<div id='instructions'>Test passes if you see no red characters.</div>
<div style="position:relative">
<div class="test" lang="hi"><div id="testdiv"><span id="testspan">हिन्दी हिन्दी हिन्दी</span><span id="testend">←</span></div></div>
<div class="ref" lang="hi"><span>हिन्दी हिन्दी हि<br/>न्दी</span><span id="refend">←</span></div>
</div>
<div id="log" style="margin-top: 10em;"></div>
<script>
var sentenceWidth = document.getElementById('testspan').offsetWidth
document.getElementById('testdiv').style.width = String(sentenceWidth - 5)+'px'
test(function() {
assert_true(document.getElementById('testend').offsetLeft === document.getElementById('refend').offsetLeft);
}, ' The text should wrap between the last two typographic letter units');
</script>
<!--
Notes:
This test is exploratory because indic conjuncts containing more than one consonant are not covered by the extended grapheme cluster definition, and therefore constitute more than one typographic unit, as defined in the CSS spec. Nevertheless, people using indic scripts expect the user agent to keep the orthographic syllable as a single unit.
-->
</body>
</html>
